Rename category .h/.m to ORKTaskResult+MORK
This allows us to only publically expose field serialization for ORKTaskResult while privately being able to do it for ORKStepResult as well
